+++ src/components/more-tabs/Changes.vue
@@ -4,18 +4,27 @@ import { changelogHTML } from '@/state/changelog';
 
 <template>
   <div class="inner-container changelog">
+    <div style="margin-bottom: 0.5em">
+      <a target="_blank" href="https://github.com/christopherwk210/tophers-archipelago-web-client/blob/main/CHANGELOG.md"><button>View on GitHub</button></a>
+    </div>
     <div class="sunken-panel" v-html="changelogHTML"></div>
   </div>
 </template>
 
 <style scoped>
+.changelog {
+  overflow: hidden;
+  display: flex;
+  flex-direction: column;
+}
+
 .inner-container {
   padding: 0 !important;
 }
 
 .sunken-panel {
   padding: 1em;
-  height: 100%;
+  flex: 1;
 }
 </style>