+++ src/views/Login.vue
@@ -119,7 +119,8 @@ async function cancel() {
   window.location.reload();
 }
 
-const hasMultipleAccounts = computed(() => localAccounts.value.length > 1);
+// const hasMultipleAccounts = computed(() => localAccounts.value.length > 1);
+const hasMultipleAccounts = computed(() => true);
 const switchAccountList = computed(() => localAccounts.value);
 
 const accountSwitcher = ref(-1);