+++ vite.config.ts
@@ -1,13 +1,16 @@
-import { fileURLToPath, URL } from 'node:url'
+import { fileURLToPath, URL } from 'node:url';
+import { defineConfig } from 'vite';
+import vue from '@vitejs/plugin-vue';
+import monacoEditorPlugin from 'vite-plugin-monaco-editor';
 
-import { defineConfig } from 'vite'
-import vue from '@vitejs/plugin-vue'
+const _monacoEditorPlugin = monacoEditorPlugin as any;
 
 // https://vite.dev/config/
 export default defineConfig({
   base: './',
   plugins: [
-    vue()
+    vue(),
+    _monacoEditorPlugin.default({})
   ],
   resolve: {
     alias: {